# swaggerAPI/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Product Management API",
    "version": "1.0.0",
    "description": "API for managing products ."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
    "/product": {
      "post": {
        "summary": " Create new product",
        "operationId": "create Product",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "productName": {
                    "type": "string",
                    "maxLength": 250
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 250
                  },
                  "price": {
                    "type": "number"
                  },
                  "quantity": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["active", "inactive"]
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["productName", "price", "quantity", "status"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "summary": "List all products",
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "description": "Filter by product name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quantityMin",
            "in": "query",
            "required": false,
            "description": "Minimum quantity for filtering",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "quantityMax",
            "in": "query",
            "required": false,
            "description": "Maximum quantity for filtering",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/product/{id}": {
      "get": {
        "summary": "Get  product by id",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the product to retrieve",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "put": {
        "summary": "Update a product",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id of the product to update",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "productName": {
                    "type": "string",
                    "maxLength": 250
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 250
                  },
                  "price": {
                    "type": "number"
                  },
                  "quantity": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["active", "inactive"]
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a product",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the product to delete",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer"
          },
          "productName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "quantity": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}